% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ylist_names.R
\name{ylist_names}
\alias{ylist_names}
\title{Search Ylist}
\usage{
ylist_names(session, query)
}
\arguments{
\item{session}{html session}

\item{query}{Japanese species name (character)}
}
\description{
Search Japanese species name, return species name.
}

